requirements.txt
agent-framework
agent-framework-ag-ui
pydantic>=2.0
sqlalchemy>=2.0
psycopg2-binary>=2.9  # or mysqlclient if you use MySQL
openai>=0.27.0  # compatible client for OpenRouter API usage
requests>=2.28
python-dotenv>=1.0  # optional, for loading env variables from .env